refactor(tagResource.js): change variable name from bgcolor to bgColor refactor(create-users.js): use faker library to generate user data The variable name bgcolor has been changed to bgColor to improve consistency with the naming conventions. The create-users.js file has been refactored to use the faker library to generate user data instead of hardcoding it. This makes it easier to generate a large number of users with realistic data for testing purposes.
feat(create-tags.js): add initial set of tags with randomized background color
This commit adds an initial set of tags with their respective name, text color, and randomized background color. The background color is randomized using the faker library's arrayElement method. This allows for a more diverse set of colors for the tags.